requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
langchain==0.1.11
pdfplumber==0.10.3
faiss-cpu==1.7.4
numpy==1.26.4
groq==0.5.0
python-dotenv==1.0.1
google-generativeai==0.3.2
pdfplumber==0.10.3
groq
